+++ R/pandoc.R
@@ -9,15 +9,21 @@
 #' between \verb{<!--pandoc} and \verb{-->}.
 #'
 #' The configuration file is a DCF file (see \code{\link{read.dcf}}). This file
-#' must contain a field named \code{format} which means the output format. The
+#' must contain a field named \code{t} which means the output format. The
 #' configurations are written in the form of \code{tag:value} and passed to
 #' Pandoc (if no value is needed, just leave it empty, e.g. the option
 #' \code{standalone} or \code{s} for short). If there are multiple output
 #' formats, write each format and relevant configurations in a block, and
 #' separate blocks with blank lines.
+#'
+#' If there are multiple records of the \code{t} field in the configuration, the
+#' input markdown file will be converted to all these formats by default, unless
+#' the \code{format} argument is specified as one single format.
 #' @param input a character vector of the Markdown filenames
 #' @param format the output format (see References); it can be a character
-#'   vector of multiple formats
+#'   vector of multiple formats; by default, it is obtained from the \code{t}
+#'   field in the configuration (if the configuration is empty or the \code{t}
+#'   field is not found, the default output format will be \code{'html'})
 #' @param config the Pandoc configuration file; if missing, it is assumed to be
 #'   a file with the same base name as the \code{input} file and an extension
 #'   \code{.pandoc} (e.g. for \file{foo.md} it looks for \file{foo.pandoc})
@@ -30,7 +36,7 @@
 #' @seealso \code{\link{read.dcf}}
 #' @export
 #' @examples system('pandoc -h') # see possible output formats
-pandoc = function(input, format = 'html', config = getOption('config.pandoc'), ext = NA) {
+pandoc = function(input, format, config = getOption('config.pandoc'), ext = NA) {
   if (Sys.which('pandoc') == '')
     stop('Please install pandoc first: http://johnmacfarlane.net/pandoc/')
   cfg = if (is.null(config)) sub_ext(input[1L], 'pandoc') else config
@@ -38,23 +44,29 @@ pandoc = function(input, format = 'html', config = getOption('config.pandoc'), e
   if (file.exists(cfg)) txt = c(txt, '', readLines(cfg, warn = FALSE))
   con = textConnection(txt); on.exit(close(con))
   cfg = read.dcf(con)
+  nms = colnames(cfg)
+  if (length(nms) && 'format' %in% nms) {
+    warning('the "format" field in the configuration must be renamed to "t"')
+    colnames(cfg)[nms == 'format'] = 't'  # for backward compatibility
+  }
+  if (missing(format)) format = pandoc_fmt(cfg)
   mapply(pandoc_one, input, format, ext, MoreArgs = list(cfg = cfg), USE.NAMES = FALSE)
 }
 # format is a scalar
 pandoc_one = function(input, format, ext, cfg) {
   cmn = NULL  # common arguments
   if (nrow(cfg) == 0L) cfg = character(0) else if (nrow(cfg) == 1L) {
-    if ('format' %in% colnames(cfg)) {
-      cfg = if (cfg[1L, 'format'] == format) drop(cfg) else NA
+    if ('t' %in% colnames(cfg)) {
+      cfg = if (cfg[1L, 't'] == format) drop(cfg) else NA
     } else {cmn = drop(cfg); cfg = NA}
   } else {
-    if (!('format' %in% colnames(cfg)))
-      stop('for a config file with multiple formats, there must be a field named "format"')
-    if (sum(idx <- is.na(cfg[, 'format'])) > 1L)
-      stop('at most one "format" field can be NA')
+    if (!('t' %in% colnames(cfg)))
+      stop('for a config file with multiple output formats, there must be a field named "t"')
+    if (sum(idx <- is.na(cfg[, 't'])) > 1L)
+      stop('at most one "t" field can be NA')
     if (sum(idx) == 1L) cmn = cfg[idx, ]
     cfg = cfg[!idx, , drop = FALSE]
-    cfg = cfg[cfg[, 'format'] == format, ]
+    cfg = cfg[cfg[, 't'] == format, ]
     if (!is.null(dim(cfg))) {
       if (nrow(cfg) > 1) stop('the output format is not unique in config')
       cfg = character(0) # nrow(cfg) == 0; format not found in cfg
@@ -65,13 +77,19 @@ pandoc_one = function(input, format, ext, cfg) {
       sub_ext(input, if (is.na(ext)) pandoc_ext(format) else ext)
     }
   })
-  cfg = cfg[setdiff(names(cfg), c('o', 'output', 'format'))]
+  cfg = cfg[setdiff(names(cfg), c('o', 'output', 't'))]
   cmd = paste('pandoc', pandoc_arg(cfg), pandoc_arg(cmn), '-f markdown',
               '-t', format, '-o', out, paste(shQuote(input), collapse = ' '))
   message('executing ', cmd)
   if (system(cmd) == 0L) out else stop('conversion failed')
 }
 
+# detect output format from config
+pandoc_fmt = function(config) {
+  fields = colnames(config)
+  if (prod(dim(config)) == 0 || !('t' %in% fields)) return('html')
+  na.omit(config[, 't'])
+}
 # infer output extension from format
 pandoc_ext = function(format) {
   if (grepl('^html', format)) return('html')